Consolidates placement of constraints
