+++ partiql-planner/src/main/kotlin/org/partiql/planner/internal/typer/PlanTyper.kt
@@ -65,9 +65,9 @@ import org.partiql.planner.internal.ir.util.PlanRewriter
 import org.partiql.planner.internal.shape.IsOrdered
 import org.partiql.planner.internal.shape.ShapeUtils
 import org.partiql.shape.Constraint
-import org.partiql.shape.Element
-import org.partiql.shape.Fields
-import org.partiql.shape.NotNull
+import org.partiql.shape.Constraint.Element
+import org.partiql.shape.Constraint.Fields
+import org.partiql.shape.Constraint.NotNull
 import org.partiql.shape.PShape
 import org.partiql.shape.PShape.Companion.allShapes
 import org.partiql.shape.PShape.Companion.allTypes
