+++ partiql-planner/src/test/kotlin/org/partiql/planner/internal/typer/PlanTyperTestsPorted.kt
@@ -31,12 +31,12 @@ import org.partiql.plugins.local.toStaticType
 import org.partiql.plugins.memory.MemoryCatalog
 import org.partiql.plugins.memory.MemoryConnector
 import org.partiql.plugins.memory.MemoryObject
-import org.partiql.shape.AnyOf
 import org.partiql.shape.Constraint
-import org.partiql.shape.Element
-import org.partiql.shape.Fields
+import org.partiql.shape.Constraint.AnyOf
+import org.partiql.shape.Constraint.Element
+import org.partiql.shape.Constraint.Fields
+import org.partiql.shape.Constraint.NotNull
 import org.partiql.shape.Meta
-import org.partiql.shape.NotNull
 import org.partiql.shape.PShape
 import org.partiql.shape.visitor.ShapePrinter
 import org.partiql.spi.BindingCase
